Modify: isolate authentication function
